requirements.txt
streamlit
agno
groq
lancedb
pandas
openpyxl
pymupdf4llm
python-docx
pillow
pypdf
duckduckgo-search
huggingface-hub
opencv-python
python-dotenv
numpy
pyarrow
sentence-transformers
requests
-e .
